flutter_tools: refactor `CreateCommand`. (#70874)
