+++ packages/flutter_tools/test/general.shard/create_config_test.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:flutter_tools/src/commands/create.dart';
+import 'package:flutter_tools/src/commands/create_base.dart';
 
 import '../src/common.dart';
